pyproject.toml
[tool.black]
line-length = 100

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests",
]

[tool.isort]
profile = "black"
line_length = 100

[tool.poetry]
name = "genotype-api"
version = "0.1.0"
description = ""
authors = ["Måns Magnusson <mans.magnusson@scilifelab.se>"]

[tool.poetry.dependencies]
python = "^3.7"
pymysql = "^0.10.1"
marshmallow_sqlalchemy = "^0.23.1"
pydantic = "^1.7.2"
fastapi = "^0.61.2"
uvicorn = "^0.12.2"
coloredlogs = "^14.0"
alembic = "^1.4.3"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
openpyxl = "^3.0.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
